MODFLOW-NWT Version 1.0.2 10/03/2011
Downloaded from http://water.usgs.gov/nrp/gwsoftware/modflow_nwt/MODFLOW-NWT_1.0.2.zip

+++ src/NWT1_xmd.f
@@ -96,7 +96,7 @@ cmi
 !     ------------------------------------------------------------------
       INTEGER lloc, istart, istop, i, n,ISTORXMD,IREDSYS, Icomb
       CHARACTER(LEN=200) line
-      REAL*4 R,RRCTOLS,EPSRNS,HCLOSEXMDDUM
+      REAL R,RRCTOLS,EPSRNS,HCLOSEXMDDUM
 !     LOCAL VARIABLES FOR XMD SOLVER
 !     ------------------------------------------------------------------
 !
